requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.5
django-cors-headers==3.4.0
django-extensions==3.0.3
django-taggit==1.5.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
ipython==7.28.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pydot==1.4.1
Pygments==2.10.0
pyparsing==2.4.7
pytz==2020.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.3.1
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
